common script

+++ update.sh
@@ -1,18 +1,23 @@
 #!/usr/bin/env bash
 
-unamestr=$(uname)
+here="$(dirname "$0")"
 
-# update yadr
-cd ~/.yadr
-git reset --hard
-git pull --rebase
-rake update
+# include common.sh
+. "$here/common.sh"
+
+# only update yadr if flagged
+if [[ $* == *-y* ]] || [[ $* == *--yadr* ]]; then
+  cd ~/.yadr
+  git reset --hard
+  git pull --rebase
+  rake update
+fi
 
 # system updates
 if [[ "$unamestr" == 'Linux' ]]; then
   sudo apt-get update
   sudo apt-get upgrade
-if [[ "$unamestr" == 'FreeBSD' ]]; then
+elif [[ "$unamestr" == 'FreeBSD' ]]; then
   brew update
   brew upgrade
 fi
